Refactor all K8s operations into their own package
